+++ changelog.md
@@ -3,6 +3,8 @@
 ## v1.17.6 / to come
 
 *  fixed nan bug in GOF, when N==0 (missing data).
+*  fixed bugs in `pre-filter`, when `--window-size` is greater than length
+*  fixed bug in `pre-filter` in pick-from-match-to mode
 
 ## v1.17.5 / 22.04.2014